# swagger.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: API de Tarefas
  description: API REST para gerenciamento de tarefas usando Java e Spring Boot. O usuário poderá cadastrar, visualizar, editar e excluir tarefas.
paths:

  /tasks:
    get:
      tags:
        - tasks
      summary: Obter todas as tarefas
      description: Retorna todas as tarefas existentes
      operationId: getAll
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Tarefa não encontrada.
        '500':
          description: Erro interno do servidor.
          
    post:
      tags:
        - tasks
      summary: Cadastrar uma nova tarefa
      description: Cria uma nova tarefa
      operationId: create
      requestBody:
        description: Criar uma nova tarefa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/TaskRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TaskRequest'
        required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Entrada inválida
        '500':
          description: Erro interno do servidor.
        '422':
          description: Validation exception
   
  /tasks/{id}:
    get:
      tags:
        - tasks
      summary: Obter tarefa pelo Id
      description: Retorna uma tarefa exsitente pelo id
      operationId: getById
      parameters:
        - name: id
          in: path
          description: ID da terafa
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: ID fornecido inválido
        '404':
          description: Tarefa não encontrada
        '422':
          description: Validation exception
          
    put:
      tags:
        - tasks
      summary: Atualizar uma tarefa existente
      description: Atualiza uma tarefa exsitente pelo id
      operationId: update
      parameters:
        - name: id
          in: path
          description: ID para tarefa que será atualizada
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Atualizar uma tarefa exsitente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/TaskRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TaskRequest'
        required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: ID fornecido inválido
        '404':
          description: Tarefa não encontrada
        '422':
          description: Validation exception
          
    delete:
      tags:
        - tasks
      summary: Excluir uma tarefa existente
      description: Exclui uma tarefa exsitente pelo id
      operationId: delete
      parameters:
        - name: id
          in: path
          description: ID da tarefa que será deletada
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Sucesso
        '400':
          description: ID fornecido inválido
        '404':
          description: Tarefa não encontrada
        '422':
          description: Validation exception
          
  /tasks/{id}/complete:     
    put:
      tags:
        - tasks
      summary: Atualizar uma tarefa como finalizada
      description: Atualiza uma tarefa exsitente como finalizada pelo id
      operationId: updateComplete
      parameters:
        - name: id
          in: path
          description: ID para tarefa que será atualizada
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: ID fornecido inválido
        '404':
          description: Tarefa não encontrada
        '422':
          description: Validation exception
          
components:
  schemas:
    TaskRequest:
      type: object
      properties:
        descricao:
          type: string
          example: Resumo de artigo
        dataCriacao:
          type: string
          format: date-time
          example: 2024-04-21
        dataLimite:
          type: string
          format: date-time
          example: 2024-04-29
        finalizada:
          type: boolean
          example: true
    TaskResponse:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 1
        - $ref: '#/components/schemas/TaskRequest'